Version 1.10 Last WP7.0 VERSION 1. Made windows bigger by removing system tray 2. Added Hebrew Translations 3. Added "Rate this program" to the main menu

+++ trunk/CrossConnect/AddNote.xaml.cs
@@ -21,17 +21,9 @@
 namespace CrossConnect
 {
     using System.Collections.Generic;
-    using System.Linq;
-    using System.Net;
     using System.Windows;
-    using System.Windows.Controls;
-    using System.Windows.Documents;
     using System.Windows.Input;
-    using System.Windows.Media;
-    using System.Windows.Media.Animation;
-    using System.Windows.Shapes;
 
-    using Microsoft.Phone.Controls;
     using Microsoft.Phone.Shell;
 
     using SwordBackend;
@@ -84,7 +76,7 @@ namespace CrossConnect
             object openWindowIndex = null;
             if (!PhoneApplicationService.Current.State.TryGetValue("openWindowIndex", out openWindowIndex))
             {
-                openWindowIndex = (int)0;
+                openWindowIndex = 0;
             }
             var state = App.openWindows[(int)openWindowIndex].state;
             //they are in reverse order again,
